+++ astsa_build/R/arma.spec.R
@@ -24,7 +24,7 @@ function(ar=0, ma=0, var.noise=1, n.freq=500, main=NULL, redundancy.tol=.1,
      red.count = 0
    for (i in 1:ar.order) {
     if ( (ar[1] == 0 && ar.order == 1) || (ma[1] == 0 && ma.order == 1) )  break
-    if(any(abs(1/z.ar[i]-1/z.ma[1:ma.order]) < redundancy.tol)) 
+    if(any(abs(1/z.ar[i]-1/z.ma) < redundancy.tol)) 
              {cat("WARNING: (Possible) Parameter Redundancy", "\n"); red.count=1; break}
    }
 
@@ -37,6 +37,7 @@ function(ar=0, ma=0, var.noise=1, n.freq=500, main=NULL, redundancy.tol=.1,
 
   spg.out <- list(freq=freq*xfreq, spec=spec)
  if(plot){
+  mane = ifelse(is.null(main), 1, 0)
   if (is.null(main)){
    if ((ar[1] == 0 && ar.order == 1) && (ma[1] == 0 && ma.order == 1)){
      main = 'White Noise'
@@ -50,11 +51,11 @@ function(ar=0, ma=0, var.noise=1, n.freq=500, main=NULL, redundancy.tol=.1,
   Xlab = ifelse(xfreq>1, paste('frequency \u00D7', xfreq), 'frequency')
    m1  = min(spec) 
    m2  = max(spec) 
-   if (red.count > 0) { 
-       ylim = c(1e-6, 10*m2)
-   } else {
    yspread =  (m2 - m1)/var.noise 
-   if (is.null(ylim) & yspread < 1) ylim = c(max(m1-2*var.noise, 0.1*var.noise), m2+2*var.noise)
+   if (mane > 0 & yspread < 1) main = "(Almost) White Noise"
+   if (is.null(ylim)) { 
+    if (yspread < 1)   ylim = c(max(m1-2*var.noise, 0.1*var.noise), m2+2*var.noise)
+    if (red.count > 0) ylim = c(1e-6, 10*m2)
    }
  tsplot(freq*xfreq, spec, xlab=Xlab, ylab=Ylab, main=main, ylim=ylim, ...)
  }